clean up: separate classes to own files + organize variables and func
