+++ Client/Frontend/Library/LibraryPanelDescriptor.swift
@@ -1,0 +1,37 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import UIKit
+
+/**
+ * Data for identifying and constructing a LibraryPanel.
+ */
+class LibraryPanelDescriptor {
+    var viewController: UIViewController?
+    var navigationController: UINavigationController?
+
+    private let profile: Profile
+    private let tabManager: TabManager
+
+    let accessibilityLabel: String
+    let accessibilityIdentifier: String
+    let panelType: LibraryPanelType
+
+    init(viewController: LibraryPanel?, profile: Profile, tabManager: TabManager, accessibilityLabel: String, accessibilityIdentifier: String, panelType: LibraryPanelType) {
+        self.viewController = viewController
+        self.profile = profile
+        self.tabManager = tabManager
+        self.accessibilityLabel = accessibilityLabel
+        self.accessibilityIdentifier = accessibilityIdentifier
+        self.panelType = panelType
+    }
+
+    // Create navigation controller if it doesn't exist
+    func setupNavigationController() {
+        guard let viewController = viewController else { return }
+
+        // Don't create a new NavigationController every time
+        navigationController = navigationController ?? ThemedNavigationController(rootViewController: viewController)
+    }
+}
